+++ lib/provider/Products.dart
@@ -12,7 +12,7 @@ import '../models/product.dart';
 import '../models/product_cart.dart';
 import '../models/product_main.dart';
 import '../models/search_detail.dart';
-import 'urls.dart';
+import '../core/constants/urls.dart';
 
 class Products with ChangeNotifier {
   List<Product> _items = [];
