+++ lib/provider/clearings.dart
@@ -8,7 +8,7 @@ import 'package:tamizshahr/models/clearing.dart';
 import 'package:tamizshahr/models/clearing_main.dart';
 
 import '../models/search_detail.dart';
-import 'urls.dart';
+import '../core/constants/urls.dart';
 
 class Clearings with ChangeNotifier {
   late String _token;
