+++ lib/provider/messages.dart
@@ -5,7 +5,7 @@ import '../models/message.dart';
 import 'package:http/http.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import 'urls.dart';
+import '../core/constants/urls.dart';
 
 class Messages with ChangeNotifier {
   List<Message> _allMessages = [];
